Updated Login and Logout functionalities

+++ core/views.py
@@ -1,6 +1,8 @@
 from django.contrib.auth.models import User
+from django.http import HttpResponse
 from django.shortcuts import redirect, render
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
 from django.db.models import Q
 from .models import Room, Topic
@@ -8,24 +10,29 @@ from .form import RoomForm
 
 
 def loginUser(request):
-    if request.method == 'POST':
+    
+    if request.user.is_authenticated:
+        return redirect('index')
+
+    if request.method == "POST":
         username = request.POST.get('username')
         password = request.POST.get('password')
 
         try:
-            user = User.objects.get(username)
+            user = User.objects.get(username=username)
         except:
-            pass
-        
+            messages.error(request, 'User does not exist')
+
         user = authenticate(request, username=username, password=password)
 
         if user != None:
             login(request, user)
             return redirect('index')
         else:
-            pass
-        
-    return render(request, 'core/login_register.html')
+            messages.error(request, "User name or password does not exist")
+
+    context = {}
+    return render(request, 'core/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
@@ -52,7 +59,7 @@ def room(request, pk):
     context = {'room':room}
     return render(request, 'core/room.html', context)
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -65,11 +72,13 @@ def createRoom(request):
     return render(request, 'core/room_form.html', context)
 
 
-
+@login_required(login_url='login')
 def editRoom(request, pk):
     room =Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!')
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -79,9 +88,11 @@ def editRoom(request, pk):
     context = {'form': form }
     return render(request, 'core/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!')
     context = {'object': room}
     if request.method == 'POST':
         room.delete()
